// src/types.ts
export type TiktakType = "naranja" | "fresa" | "frutas" | "menta";

export interface Answers {
  q1: TiktakType | null;
  q2: TiktakType | null;
  q3: TiktakType | null;
  q4: TiktakType | null;
  q5: TiktakType | null;
  q6: TiktakType | null;
  q7: TiktakType | null;
  q8: TiktakType | null;
  q9: TiktakType | null;
  q10: TiktakType | null;
}